Move SchemaBuilder to DataViewSchema.Builder.
