+++ src/Microsoft.Data.DataView/DataViewSchema.cs
@@ -352,7 +352,7 @@ namespace Microsoft.Data.DataView
                 /// </summary>
                 public Metadata GetMetadata()
                 {
-                    var builder = new SchemaBuilder();
+                    var builder = new DataViewSchema.Builder();
                     foreach (var item in _items)
                         builder.AddColumn(item.Name, item.Type, item.Metadata);
                     return new Metadata(builder.GetSchema(), _items.Select(x => x.Getter).ToArray());
@@ -373,10 +373,78 @@ namespace Microsoft.Data.DataView
         }
 
         /// <summary>
-        /// This constructor should only be called by <see cref="SchemaBuilder"/>.
+        /// The class that incrementally builds a <see cref="DataViewSchema"/>.
+        /// </summary>
+        public sealed class Builder
+        {
+            private readonly List<(string Name, DataViewType Type, Metadata Metadata)> _items;
+
+            /// <summary>
+            /// Create a new instance of <see cref="Builder"/>.
+            /// </summary>
+            public Builder()
+            {
+                _items = new List<(string Name, DataViewType Type, Metadata Metadata)>();
+            }
+
+            /// <summary>
+            /// Add one column to the schema being built.
+            /// </summary>
+            /// <param name="name">The column name.</param>
+            /// <param name="type">The column type.</param>
+            /// <param name="metadata">The column metadata.</param>
+            public void AddColumn(string name, DataViewType type, Metadata metadata = null)
+            {
+                if (string.IsNullOrEmpty(name))
+                    throw new ArgumentNullException(nameof(name));
+                if (type == null)
+                    throw new ArgumentNullException(nameof(type));
+
+                _items.Add((name, type, metadata));
+            }
+
+            /// <summary>
+            /// Add multiple existing columns to the schema being built.
+            /// </summary>
+            /// <param name="source">Columns to add.</param>
+            public void AddColumns(IEnumerable<Column> source)
+            {
+                foreach (var column in source)
+                    AddColumn(column.Name, column.Type, column.Metadata);
+            }
+
+            /// <summary>
+            /// Add multiple existing columns to the schema being built.
+            /// </summary>
+            /// <param name="source">Columns to add.</param>
+            public void AddColumns(IEnumerable<DetachedColumn> source)
+            {
+                foreach (var column in source)
+                    AddColumn(column.Name, column.Type, column.Metadata);
+            }
+
+            /// <summary>
+            /// Generate the final <see cref="DataViewSchema"/>.
+            /// </summary>
+            public DataViewSchema GetSchema()
+            {
+                var nameMap = new Dictionary<string, int>();
+                for (int i = 0; i < _items.Count; i++)
+                    nameMap[_items[i].Name] = i;
+
+                var columns = new Column[_items.Count];
+                for (int i = 0; i < columns.Length; i++)
+                    columns[i] = new Column(_items[i].Name, i, nameMap[_items[i].Name] != i, _items[i].Type, _items[i].Metadata);
+
+                return new DataViewSchema(columns);
+            }
+        }
+
+        /// <summary>
+        /// This constructor should only be called by <see cref="Builder"/>.
         /// </summary>
         /// <param name="columns">The input columns. The constructed instance takes ownership of the array.</param>
-        internal DataViewSchema(Column[] columns)
+        private DataViewSchema(Column[] columns)
         {
             if (columns == null)
                 throw new ArgumentNullException(nameof(columns));
