+++ src/Microsoft.ML.Core/Data/SchemaExtensions.cs
@@ -12,7 +12,7 @@ namespace Microsoft.ML.Data
     {
         public static DataViewSchema MakeSchema(IEnumerable<DataViewSchema.DetachedColumn> columns)
         {
-            var builder = new SchemaBuilder();
+            var builder = new DataViewSchema.Builder();
             builder.AddColumns(columns);
             return builder.GetSchema();
         }
