+++ src/Microsoft.ML.Data/DataView/ZipBinding.cs
@@ -35,7 +35,7 @@ namespace Microsoft.ML.Data
                 _cumulativeColCounts[i + 1] = _cumulativeColCounts[i] + schema.Count;
             }
 
-            var schemaBuilder = new SchemaBuilder();
+            var schemaBuilder = new DataViewSchema.Builder();
             foreach (var sourceSchema in sources)
                 schemaBuilder.AddColumns(sourceSchema);
             OutputSchema = schemaBuilder.GetSchema();
