+++ src/Microsoft.ML.Parquet/ParquetLoader.cs
@@ -317,7 +317,7 @@ namespace Microsoft.ML.Data
         {
             Contracts.AssertValue(ectx);
             Contracts.AssertValue(cols);
-            var builder = new SchemaBuilder();
+            var builder = new DataViewSchema.Builder();
             builder.AddColumns(cols.Select(c => new Microsoft.Data.DataView.DataViewSchema.DetachedColumn(c.Name, c.ColType, null)));
             return builder.GetSchema();
         }
