+++ src/Microsoft.ML.TensorFlow/TensorFlow/TensorflowUtils.cs
@@ -30,7 +30,7 @@ namespace Microsoft.ML.Transforms.TensorFlow
 
         internal static DataViewSchema GetModelSchema(IExceptionContext ectx, TFGraph graph, string opType = null)
         {
-            var schemaBuilder = new SchemaBuilder();
+            var schemaBuilder = new DataViewSchema.Builder();
             foreach (var op in graph)
             {
                 if (opType != null && opType != op.OpType)
