+++ test/Microsoft.ML.Tests/FakeSchemaTest.cs
@@ -23,7 +23,7 @@ namespace Microsoft.ML.Tests
         {
             var metadataBuilder = new DataViewSchema.Metadata.Builder();
             metadataBuilder.Add("M", NumberDataViewType.Single, (ref float v) => v = 484f);
-            var schemaBuilder = new SchemaBuilder();
+            var schemaBuilder = new DataViewSchema.Builder();
             schemaBuilder.AddColumn("A", new VectorType(NumberDataViewType.Single, 94));
             schemaBuilder.AddColumn("B", new KeyType(typeof(uint), 17));
             schemaBuilder.AddColumn("C", NumberDataViewType.Int32, metadataBuilder.GetMetadata());
